pyproject.toml
[project]
name = "deep-research-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.5.0",
    "litellm>=1.73.1",
    "openai>=1.91.0",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
]
